+++ apps/citizen-func/src/functions/GetMessages/__tests__/handler.test.ts
@@ -923,7 +923,7 @@ describe("GetMessagesHandler |> Fallback |> Enrichment", () => {
         tag: TagEnumPN.PN,
         ...aPnThirdPartyData,
       },
-      has_attachments: false,
+      has_attachments: true,
       has_precondition: true,
       has_remote_content: true,
       is_archived: false,