+++ infra/resources/_modules/mongo/payments.tf
@@ -16,7 +16,7 @@ module "payments_cosmos_account" {
   is_virtual_network_filter_enabled = false
 
   main_geo_location_location       = var.environment.location
-  main_geo_location_zone_redundant = false
+  main_geo_location_zone_redundant = true
   additional_geo_locations = [{
     location          = var.secondary_geo_location
     failover_priority = 1