+++ apps/services-func/CHANGELOG.md
@@ -1,5 +1,19 @@
 # services-func
 
+## 2.1.4
+
+### Patch Changes
+
+- d13e694: Remove legal data from io-messages-common-legacy message-content model
+- Updated dependencies [d13e694]
+  - io-messages-common-legacy@0.1.1
+
+## 2.1.3
+
+### Patch Changes
+
+- 74207e6: Splitting config and removing connection string for production env
+
 ## 2.1.2
 
 ### Patch Changes